Added missing passwords functions to example app

+++ sdk/src/main/java/com/stytch/sdk/network/StytchApi.kt
@@ -154,13 +154,13 @@ internal object StytchApi {
         suspend fun authenticate(
             email: String,
             password: String,
-            sessionDurationMinutes: Int,
+            sessionDurationMinutes: UInt,
         ): StytchResult<AuthData> = safeApiCall {
             apiService.authenticateWithPasswords(
                 StytchRequests.Passwords.AuthenticateRequest(
                     email,
                     password,
-                    sessionDurationMinutes
+                    sessionDurationMinutes.toInt()
                 )
             )
         }
@@ -168,13 +168,13 @@ internal object StytchApi {
         suspend fun create(
             email: String,
             password: String,
-            sessionDurationMinutes: Int,
+            sessionDurationMinutes: UInt,
         ): StytchResult<CreateResponse> = safeApiCall {
             apiService.passwords(
                 StytchRequests.Passwords.CreateRequest(
                     email,
                     password,
-                    sessionDurationMinutes
+                    sessionDurationMinutes.toInt()
                 )
             )
         }
@@ -204,14 +204,14 @@ internal object StytchApi {
         suspend fun resetByEmail(
             token: String,
             password: String,
-            sessionDurationMinutes: Int,
+            sessionDurationMinutes: UInt,
             codeVerifier: String,
         ): StytchResult<AuthData> = safeApiCall {
             apiService.resetByEmail(
                 StytchRequests.Passwords.RestByEmailRequest(
                     token,
                     password,
-                    sessionDurationMinutes,
+                    sessionDurationMinutes.toInt(),
                     codeVerifier
                 )
             )
